Completed Module-7 finally. The error mentioned int he previous commit was found.   I nested the services.AddScoped<ICityInfoRepository CityInfoRepository>(); inside the services.AddDbContext<CityInfoContext>... options.   Rookie mistake...
There is still an error when the application is started in Debug.   To Explain:
In the Program Class, there is a call to "context.Database.EnsureDeleted();" followed by a call to "context.Database.Migrate();"   Just after the database is deleted, the Exception is-
Exception thrown: 'System.Data.SqlClient.SqlException' in System.Data.SqlClient.dll-
But if you wait a few moments, it seems to cure itself.   I might have something wrong in the migration files, but I have no idea what.